+++ raphael-solver/src/macro_solver/search_queue.rs
@@ -12,7 +12,7 @@ use super::pareto_front::ParetoFront;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct SearchScore {
-    pub quality_upper_bound: u32,
+    pub quality_upper_bound: u16,
     pub steps_lower_bound: u8,
     pub duration_lower_bound: u8,
     pub current_steps: u8,
@@ -29,7 +29,7 @@ impl SearchScore {
     };
 
     pub const MAX: Self = Self {
-        quality_upper_bound: u32::MAX,
+        quality_upper_bound: u16::MAX,
         steps_lower_bound: 0,
         duration_lower_bound: 0,
         current_steps: 0,
